initial attempt on refactoring
